Initial checkin
Still working on the extruder feed rate. It was 87 - result is wierd little bumps on each level. Most likely its because of the extruder

the heat bed tuning was performed with the marlin autotune. the heat bed is not insulated at all. I'm only going to use the 6x6 plate long enough to print the 8x8 extenders. ( well, and whenever panucat gets the heatbed shipped! )

no pid tuning has been performed on the hot end